Preserve control characters
If a control character like `\u0002` appears in the XML it is preserved by the REXML parser, but Nokogiri parser bails out with an incomplete XML. Note that scrubbing the string does not help in this case since this is a valid Unicode character, but it is invalid in XML 1.0.

To handle this we extract the character from the error message. For parsing to continue we must also tell Nokogiri to recover from errors.